rename a package and struct
